integrations/triton/requirements.txt
pika==1.3.1
numpy==1.24.3
tritonclient[all]==2.34.0
pyyaml==6.0
redis==4.5.5
aiohttp==3.8.4
jsonschema==4.17.3
